+++ s/server/examples/server.ts
@@ -1,9 +1,8 @@
 
 import {Server} from "../server.js"
-import {ExClientside} from "./types.js"
 import {LoggerTap} from "../../core/taps/logger.js"
+import {exampleRpc, exampleAccepter} from "./logic.js"
 import {route} from "../../transports/http/parts/routing.js"
-import {exampleHttpRpc, exampleWsServersideRpc} from "./rpcs.js"
 import {respond} from "../../transports/http/parts/responding.js"
 
 export const port = 8000
@@ -12,14 +11,9 @@ export const logger = new LoggerTap()
 await new Server({
 	tap: logger,
 	cors: {origins: "*"},
-	rpc: exampleHttpRpc,
-	websocket: Server.websocket<ExClientside>(async _connection => ({
-		rpc: exampleWsServersideRpc,
-		disconnected: () => {},
-	})),
-	routes: [
-		route.get("/lol", respond.text("rofl")),
-	],
+	routes: [route.get("/lol", respond.text("rofl"))],
+	rpc: exampleRpc,
+	websocket: exampleAccepter,
 }).listen(port)
 
 await logger.log(`renraku server :${port}...`)