+++ s/server/examples/types.ts
@@ -1,9 +1,0 @@
-
-export type ExServerside = {
-	now(): Promise<number>
-}
-
-export type ExClientside = {
-	sum(a: number, b: number): Promise<number>
-}
-